небольшие фиксы оформления
