+++ CryptoWallet/Screens/MainTabs/CryptoList/CryptoSingle/SingleCryptoViewModelProtocol.swift
@@ -1,0 +1,13 @@
+//
+//  SingleCryptoViewModelProtocol.swift
+//  CryptoWallet
+//
+//  Created by Ilya Kuznetsov on 11.04.2025.
+//
+
+import Foundation
+
+
+protocol SingleCryptoViewModelProtocol {
+    var coin: CryptoCurrency { get }
+}
